Update WebApp Client & Config WPF

+++ ChatProgramm-WebApp/login/script.js
@@ -1,7 +1,8 @@
 //Autor: Stefan Rautner
 
-//Variablen definieren
-let uID = "";
+//Globale Variablen definieren (UrlZumSpringBootServer & urlParameter)
+localStorage.setItem('urlToSpringBootServer', 'http://localhost:8080/tinyWhatsApp');
+localStorage.setItem('urlParameter', window.location.search);
 
 //Zu RegisterForm wechseln
 function changeToLogin() {
@@ -39,9 +40,6 @@ function changeToDelete() {
     document.getElementById("deleteForm").style.display = "block";
 }
 
-// URL zur MongoDB Datenbank definieren
-const urlToSpringBoot = 'http://localhost:8080/tinyWhatsApp';
-
 //Passwort hashen
 async function hash(password) {
     const encoder = new TextEncoder();
@@ -54,25 +52,31 @@ async function hash(password) {
 //Benutzer überprüfen
 async function checkUserExistence(event) {
     try {
+        const username = document.getElementById("usernameLogin").value;
+        const password = document.getElementById("passwordLogin").value;
         event.preventDefault();
-        const response = await fetch(`${urlToSpringBoot}/checkUser`, {
-            method: 'POST',
-            body: JSON.stringify({
-                'username': document.getElementById("usernameLogin").value,
-                'password': await hash(document.getElementById("passwordLogin").value)
-            })
-        });
+        if(username !== null && username !== "" && password !== null && password !== "") {
+            const response = await fetch(`${localStorage.getItem('urlToSpringBootServer')}/checkUser`, {
+                method: 'POST',
+                body: JSON.stringify({
+                    'username': username,
+                    'password': await hash(password)
+                })
+            });
 
-        //Form leeren
-        document.getElementById("usernameLogin").value = "";
-        document.getElementById("passwordLogin").value = "";
+            //Form leeren
+            document.getElementById("usernameLogin").value = "";
+            document.getElementById("passwordLogin").value = "";
 
-        const data = await response.text();
-        if (data !== null && data !== "") {
-            uID = data.userID;
-            window.location.href = '../home/home.html' + window.urlParameter;
+            const data = await response.text();
+            if (data !== null && data !== "") {
+                localStorage.setItem('userID', data.userID);
+                window.location.href = '../home/home.html' + localStorage.getItem('urlParameter');
+            } else {
+                alert("Anmeldung fehlgeschlagen");
+            }
         } else {
-            alert("Anmeldung fehlgeschlagen");
+            alert("Bitte geben Sie einen Benutzernamen und Passwort ein");
         }
     } catch (error) {
         console.error(error);
@@ -83,26 +87,32 @@ async function checkUserExistence(event) {
 //Benutzer hinzufügen
 async function createNewUser(event) {
     try {
+        const username = document.getElementById("usernameRegister").value;
+        const password = document.getElementById("passwordRegister").value;
         event.preventDefault();
-        const response = await fetch(`${urlToSpringBoot}/newUser`, {
-            method: 'POST',
-            body: JSON.stringify({
-                'username': document.getElementById("usernameRegister").value,
-                'password': await hash(document.getElementById("passwordRegister").value)
-            })
-        });
+        if(username !== null && username !== "" && password !== null && password !== "") {
+            const response = await fetch(`${localStorage.getItem('urlToSpringBootServer')}/newUser`, {
+                method: 'POST',
+                body: JSON.stringify({
+                    'username': username,
+                    'password': await hash(password)
+                })
+            });
 
-        //Form leeren & zu Login wechseln
-        document.getElementById("usernameRegister").value = "";
-        document.getElementById("passwordRegister").value = "";
-        changeToLogin();
+            //Form leeren & zu Login wechseln
+            document.getElementById("usernameRegister").value = "";
+            document.getElementById("passwordRegister").value = "";
+            changeToLogin();
 
-        const data = await response.text();
-        if (data !== null && data !== "") {
-            uID = data.userID;
-            window.location.href = '../home/home.html' + window.urlParameter;
+            const data = await response.text();
+            if (data !== null && data !== "") {
+                localStorage.setItem('userID', data.userID);
+                window.location.href = '../home/home.html' + localStorage.getItem('urlParameter');
+            } else {
+                alert("Registrierung fehlgeschlagen");
+            }
         } else {
-            alert("Registrierung fehlgeschlagen");
+            alert("Bitte geben Sie einen Benutzernamen und Passwort ein");
         }
     } catch (error) {
         console.error(error);
@@ -112,26 +122,32 @@ async function createNewUser(event) {
 //Benutzer aktualisieren/updaten
 async function updateUser(event) {
     try {
+        const username = document.getElementById("usernamePasswordLost").value;
+        const password = document.getElementById("passwordPasswordLost").value;
         event.preventDefault();
-        const response = await fetch(`${urlToSpringBoot}/updateUser`, {
-            method: 'PUT',
-            body: JSON.stringify({
-                'username': document.getElementById("usernamePasswordLost").value,
-                'password': await hash(document.getElementById("passwordPasswordLost").value)
-            })
-        });
+        if(username !== null && username !== "" && password !== null && password !== "") {
+            const response = await fetch(`${localStorage.getItem('urlToSpringBootServer')}/updateUser`, {
+                method: 'PUT',
+                body: JSON.stringify({
+                    'username': username,
+                    'password': await hash(password)
+                })
+            });
 
-        //Form leeren & zu Login wechseln
-        document.getElementById("usernamePasswordLost").value = "";
-        document.getElementById("passwordPasswordLost").value = "";
-        changeToLogin();
+            //Form leeren & zu Login wechseln
+            document.getElementById("usernamePasswordLost").value = "";
+            document.getElementById("passwordPasswordLost").value = "";
+            changeToLogin();
 
-        const data = await response.text();
-        if (data !== null && data !== "") {
-            uID = data.userID;
-            window.location.href = '../home/home.html' + window.urlParameter;
+            const data = await response.text();
+            if (data !== null && data !== "") {
+                localStorage.setItem('userID', data.userID);
+                window.location.href = '../home/home.html' + localStorage.getItem('urlParameter');
+            } else {
+                alert("Passwort aktualisieren fehlgeschlagen");
+            }
         } else {
-            alert("Passwort aktualisieren fehlgeschlagen");
+            alert("Bitte geben Sie einen Benutzernamen und Passwort ein");
         }
     } catch (error) {
         console.error(error);
@@ -141,27 +157,28 @@ async function updateUser(event) {
 //Benutzer aktualisieren/updaten
 async function deleteUser(event) {
     try {
+        const username = document.getElementById("usernameDelete").value;
+        const password = document.getElementById("passwordDelete").value;
         event.preventDefault();
-        const response = await fetch(`${urlToSpringBoot}/deleteUser`, {
-            method: 'DELETE',
-            body: JSON.stringify({
-                'username': document.getElementById("usernameDelete").value,
-                'password': await hash(document.getElementById("passwordDelete").value)
-            })
-        });
+        if(username !== null && username !== "" && password !== null && password !== "") {
+            const response = await fetch(`${localStorage.getItem('urlToSpringBootServer')}/deleteUser`, {
+                method: 'DELETE',
+                body: JSON.stringify({
+                    'username': username,
+                    'password': await hash(password)
+                })
+            });
 
-        //Form leeren & zu Login wechseln
-        document.getElementById("usernameDelete").value = "";
-        document.getElementById("passwordDelete").value = "";
-        changeToLogin();
+            //Form leeren & zu Login wechseln
+            document.getElementById("usernameDelete").value = "";
+            document.getElementById("passwordDelete").value = "";
+            changeToLogin();
 
-        alert(await response.text());
+            alert(await response.text());
+        } else {
+            alert("Bitte geben Sie einen Benutzernamen und Passwort ein");
+        }
     } catch (error) {
         console.error(error);
     }
 }
-
-//Variable für andere Scripte verfügbar machen
-window.userID = uID;
-window.urlToSpringBoot = urlToSpringBoot;
-window.urlParameter = window.location.search;
